+++ packages/apollo-angular/testing/src/module.ts
@@ -7,8 +7,14 @@ import {NgModule, InjectionToken, Inject, Optional} from '@angular/core';
 import {ApolloTestingController} from './controller';
 import {ApolloTestingBackend} from './backend';
 
+export type NamedCaches = Record<string, ApolloCache<any> | undefined | null>;
+
 export const APOLLO_TESTING_CACHE = new InjectionToken<ApolloCache<any>>(
   'apollo-angular/testing cache',
+);
+
+export const APOLLO_TESTING_NAMED_CACHE = new InjectionToken<NamedCaches>(
+  'apollo-angular/testing named cache',
 );
 
 @NgModule({
@@ -25,14 +31,29 @@ export class ApolloTestingModule {
     @Optional()
     @Inject(APOLLO_TESTING_CACHE)
     cache?: ApolloCache<any>,
+    @Optional()
+    @Inject(APOLLO_TESTING_NAMED_CACHE)
+    namedCaches?: NamedCaches,
   ) {
-    apollo.create({
-      link: new ApolloLink(operation => backend.handle(operation)),
-      cache:
-        cache ||
-        new InMemoryCache({
-          addTypename: false,
-        }),
-    });
+    function createOptions(c?: ApolloCache<any> | null) {
+      return {
+        link: new ApolloLink(operation => backend.handle(operation)),
+        cache:
+          c ||
+          new InMemoryCache({
+            addTypename: false,
+          }),
+      };
+    }
+
+    apollo.create(createOptions(cache));
+
+    if (namedCaches && typeof namedCaches === 'object') {
+      for (const name in namedCaches) {
+        if (namedCaches.hasOwnProperty(name)) {
+          apollo.createNamed(name, createOptions(namedCaches[name]));
+        }
+      }
+    }
   }
 }